Feat(ActionChooser): Added Red-Border for min-app-version incompatibility to Wallpaper and Icon Page. Moved Version check functions into Preview.py and StorePage.py
